Backend: Added auth middleware

+++ todo-backend/src/todos/todos.controller.ts
@@ -6,14 +6,14 @@ import { TodosService } from './todos.service';
 export class TodosController {
   constructor(private readonly todosService: TodosService) {}
 
-  @Get('/all')
+  @Get('/')
   async getAllTodos(
-    @Req() request: Request,
+    @Req() request: Request & { user: { userId: string } },
     @Res() response: Response,
   ): Promise<any> {
     try {
       const todos = await this.todosService.getTodoByUser(
-        request.body.userId as string,
+        request.user.userId as string,
       );
       return response.status(200).json(todos);
     } catch (error) {
@@ -21,9 +21,22 @@ export class TodosController {
     }
   }
 
+  @Get('/:id')
+  async getTodoById(
+    @Req() request: Request,
+    @Res() response: Response,
+  ): Promise<any> {
+    try {
+      const todo = await this.todosService.getTodoById(request.params.id);
+      return response.status(200).json(todo);
+    } catch (error) {
+      return response.status(500).json({ message: 'Todo does not exist!' });
+    }
+  }
+
   @Post('/')
   async createTodo(
-    @Req() request: Request,
+    @Req() request: Request & { user: { userId: string } },
     @Res() response: Response,
   ): Promise<any> {
     try {
@@ -32,7 +45,7 @@ export class TodosController {
         description: request.body.description,
         priority: request.body.priority,
         date: request.body.date,
-        userId: request.body.userId,
+        userId: request.user.userId as string,
       });
       return response.status(201).json(todo);
     } catch (error) {
@@ -42,12 +55,17 @@ export class TodosController {
 
   @Put('/:id')
   async updateTodo(
-    @Req() request: Request,
+    @Req() request: Request & { user: { userId: string } },
     @Res() response: Response,
   ): Promise<any> {
     try {
       const todoId = request.params.id;
+      const unUpdatedTodo = await this.todosService.getTodoById(todoId);
       const updateFields: any = {};
+
+      if (unUpdatedTodo.userId !== request.user.userId) {
+        return response.status(401).json({ message: 'Unauthorized' });
+      }
 
       if (request.body.title !== undefined) {
         updateFields.title = request.body.title;
@@ -74,10 +92,17 @@ export class TodosController {
 
   @Delete('/:id')
   async deleteTodo(
-    @Req() request: Request,
+    @Req() request: Request & { user: { userId: string } },
     @Res() response: Response,
   ): Promise<any> {
     try {
+      const unDeletedTodo = await this.todosService.getTodoById(
+        request.params.id,
+      );
+      if (unDeletedTodo.userId !== request.user.userId) {
+        return response.status(401).json({ message: 'Unauthorized' });
+      }
+
       const todo = await this.todosService.deleteTodoById(request.params.id);
       return response.status(200).json(todo);
     } catch (error) {
